backend/app/tests/integration_test_auth.py
from __future__ import annotations

import os
import tempfile
from unittest.mock import patch

import pytest
from fastapi.testclient import TestClient
from sqlmodel import Session, SQLModel, create_engine

from backend.app.api.routes.tasks import utc_now
from backend.app.core.config import Settings, get_settings
from backend.app.db.models.task import Task
from backend.app.main import app
from backend.app.db.models.user import User


@pytest.fixture(name="engine")
def engine_fixture():
    """Create a temporary SQLite database for testing."""
    _, db_path = tempfile.mkstemp(suffix=".db")
    try:
        database_url = f"sqlite:///{db_path}"
        engine = create_engine(database_url)
        yield engine
    finally:
        os.unlink(db_path)


@pytest.fixture(name="session")
def session_fixture(engine):
    """Create a test database session."""
    SQLModel.metadata.create_all(bind=engine)
    with Session(engine) as session:
        yield session


@pytest.fixture(name="client")
def client_fixture(engine):
    """Create a test client with mocked settings."""
    def get_test_settings():
        return Settings(
            database_url=f"{engine.url}",
            better_auth_secret="test-secret-key-for-dev-only",
            app_name="tasks-api-test"
        )

    app.dependency_overrides[get_settings] = get_test_settings
    client = TestClient(app)
    yield client
    app.dependency_overrides.clear()


def test_create_task_without_auth_header_fails(client: TestClient):
    """Test that creating a task without authentication header fails."""
    user_id = "test-user-id"
    response = client.post(f"/api/{user_id}/tasks", json={
        "title": "Test Task",
        "description": "Test Description"
    })
    assert response.status_code == 401
    assert response.json()["error"]["code"] == "AUTH_UNAUTHORIZED"


def test_create_task_with_invalid_jwt_fails(client: TestClient):
    """Test that creating a task with invalid JWT fails."""
    user_id = "test-user-id"
    headers = {"Authorization": "Bearer invalid-token"}
    response = client.post(f"/api/{user_id}/tasks", json={
        "title": "Test Task",
        "description": "Test Description"
    }, headers=headers)
    assert response.status_code == 401
    assert response.json()["error"]["code"] == "AUTH_UNAUTHORIZED"


def test_task_ownership_enforcement(client: TestClient, session: Session):
    """Test that users can only access their own tasks."""
    # Create a task for user1
    user1_id = "user1"
    user2_id = "user2"

    task = Task(
        user_id=user1_id,
        title="User1's Task",
        description="Test Description",
        completed=False,
    )
    session.add(task)
    session.commit()
    session.refresh(task)

    # Mock a valid JWT for user2
    with patch("backend.app.core.security.jwt.decode") as mock_decode:
        mock_decode.return_value = {"sub": user2_id}

        # Try to access user1's task as user2 (should fail)
        headers = {"Authorization": "Bearer valid-token-for-user2"}
        response = client.get(f"/api/{user2_id}/tasks/{task.id}", headers=headers)
        assert response.status_code == 404  # Task not found because it belongs to user1

        # Try to access user1's task as user1 (should succeed)
        with patch("backend.app.core.security.jwt.decode") as mock_decode_user1:
            mock_decode_user1.return_value = {"sub": user1_id}
            headers_user1 = {"Authorization": "Bearer valid-token-for-user1"}
            response = client.get(f"/api/{user1_id}/tasks/{task.id}", headers=headers_user1)
            assert response.status_code == 200
            assert response.json()["id"] == task.id
            assert response.json()["user_id"] == user1_id


def test_jwt_verification_utility():
    """Test JWT verification utility functionality."""
    from backend.app.core.security import verify_jwt_and_get_token
    from backend.app.core.errors import http_error

    # Test with valid token
    import jwt
    from datetime import datetime, timedelta, timezone

    secret = "test-secret"
    user_id = "test-user-id"
    exp_time = datetime.now(timezone.utc) + timedelta(hours=1)
    token_payload = {
        "sub": user_id,
        "exp": exp_time.timestamp(),
        "iat": datetime.now(timezone.utc).timestamp()
    }
    token = jwt.encode(token_payload, secret, algorithm="HS256")

    # Mock the dependencies
    from backend.app.core.config import Settings
    settings = Settings(
        database_url="sqlite:///:memory:",
        better_auth_secret=secret,
        app_name="test"
    )

    # Create mock request with valid authorization header
    from starlette.datastructures import Headers
    from backend.app.core.security import AuthToken

    # Since we can't easily mock the full FastAPI dependency injection here,
    # we'll just verify that the core JWT functionality works as expected
    decoded = jwt.decode(token, secret, algorithms=["HS256"], options={"require": ["sub"]})
    assert decoded["sub"] == user_id


def test_cross_user_access_prevention(client: TestClient, session: Session):
    """Test that cross-user task access is prevented."""
    # Create tasks for two different users
    user1_id = "user1"
    user2_id = "user2"

    task1 = Task(
        user_id=user1_id,
        title="User1's Task",
        description="Test Description",
        completed=False,
    )
    task2 = Task(
        user_id=user2_id,
        title="User2's Task",
        description="Test Description",
        completed=False,
    )
    session.add(task1)
    session.add(task2)
    session.commit()

    # Mock JWT for user1
    with patch("backend.app.core.security.jwt.decode") as mock_decode:
        mock_decode.return_value = {"sub": user1_id}

        headers = {"Authorization": "Bearer valid-token-for-user1"}

        # User1 should be able to list their own tasks
        response = client.get(f"/api/{user1_id}/tasks", headers=headers)
        assert response.status_code == 200
        user1_tasks = response.json()
        assert len([t for t in user1_tasks if t["id"] == task1.id]) == 1
        assert len([t for t in user1_tasks if t["id"] == task2.id]) == 0  # Should not see user2's task

        # User1 should not be able to access user2's task
        response = client.get(f"/api/{user1_id}/tasks/{task2.id}", headers=headers)
        assert response.status_code == 404  # Not found because it belongs to user2