backend/pyproject.toml
[project]
name = "tasks-backend"
version = "0.1.0"
description = "Backend API for user-scoped tasks"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.22",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.7.0",
  "pyjwt>=2.9.0",
  "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
